[CHORE] workflow controller created for workflow related actions
