+++ server/src/controllers/index.ts
@@ -1,2 +1,4 @@
 export * from "./github.controller";
 export * from "./webhook.controller";
+export * from "./workflow.controller";
+export * from "./logincontroller";
