+++ server/src/routes/github.router.ts
@@ -1,10 +1,9 @@
 import express from "express";
-import { register_webhook, repositories , get_workflows} from "../controllers";
+import { register_webhook, repositories } from "../controllers";
 
 const githubRouter = express.Router();
 
 githubRouter.post("/register/webhook", register_webhook);
 githubRouter.post("/user/repositories", repositories);
-githubRouter.get("/workflows/:user_id", get_workflows);
 
 export default githubRouter;
