+++ server/src/routes/index.ts
@@ -1,2 +1,4 @@
 export * from "./auth.router";
 export * from "./github.router";
+export * from "./workflow.router";
+export * from "./webhook.router";
